+++ lib/paper_trail/version_concern.rb
@@ -7,8 +7,7 @@ module PaperTrail
     included do
       belongs_to :item, :polymorphic => true
       validates_presence_of :event
-      attr_accessible :item_type, :item_id, :event, :whodunnit, :object, :object_changes if PaperTrail.active_record_protected_attributes?
-
+      attr_accessible :item_type, :item_id, :event, :whodunnit, :object, :object_changes, :created_at if PaperTrail.active_record_protected_attributes?
       after_create :enforce_version_limit!
     end
 
@@ -37,35 +36,50 @@ module PaperTrail
       # `timestamp_arg` receives `true`
       def subsequent(obj, timestamp_arg = false)
         if timestamp_arg != true && self.primary_key_is_int?
-          return where("#{table_name}.#{self.primary_key} > ?", obj).order("#{table_name}.#{self.primary_key} ASC")
+          return where(arel_table[primary_key].gt(obj.id)).order(arel_table[primary_key].asc)
         end
 
         obj = obj.send(PaperTrail.timestamp_field) if obj.is_a?(self)
-        where("#{table_name}.#{PaperTrail.timestamp_field} > ?", obj).order(self.timestamp_sort_order)
+        where(arel_table[PaperTrail.timestamp_field].gt(obj)).order(self.timestamp_sort_order)
       end
 
       def preceding(obj, timestamp_arg = false)
         if timestamp_arg != true && self.primary_key_is_int?
-          return where("#{table_name}.#{self.primary_key} < ?", obj).order("#{table_name}.#{self.primary_key} DESC")
+          return where(arel_table[primary_key].lt(obj.id)).order(arel_table[primary_key].desc)
         end
 
         obj = obj.send(PaperTrail.timestamp_field) if obj.is_a?(self)
-        where("#{table_name}.#{PaperTrail.timestamp_field} < ?", obj).order(self.timestamp_sort_order('DESC'))
+        where(arel_table[PaperTrail.timestamp_field].lt(obj)).order(self.timestamp_sort_order('desc'))
       end
 
 
       def between(start_time, end_time)
-        where("#{table_name}.#{PaperTrail.timestamp_field} > ? AND #{table_name}.#{PaperTrail.timestamp_field} < ?",
-          start_time, end_time).order(self.timestamp_sort_order)
+        where(
+          arel_table[PaperTrail.timestamp_field].gt(start_time).
+          and(arel_table[PaperTrail.timestamp_field].lt(end_time))
+        ).order(self.timestamp_sort_order)
       end
 
       # defaults to using the primary key as the secondary sort order if possible
-      def timestamp_sort_order(order = 'ASC')
-        if self.primary_key_is_int?
-          "#{table_name}.#{PaperTrail.timestamp_field} #{order}, #{table_name}.#{self.primary_key} #{order}"
-        else
-          "#{table_name}.#{PaperTrail.timestamp_field} #{order}"
+      def timestamp_sort_order(direction = 'asc')
+        [arel_table[PaperTrail.timestamp_field].send(direction.downcase)].tap do |array|
+          array << arel_table[primary_key].send(direction.downcase) if self.primary_key_is_int?
         end
+      end
+
+      # Performs an attribute search on the serialized object by invoking the
+      # identically-named method in the serializer being used.
+      def where_object(args = {})
+        raise ArgumentError, 'expected to receive a Hash' unless args.is_a?(Hash)
+        arel_field = arel_table[:object]
+
+        where_conditions = args.map do |field, value|
+          PaperTrail.serializer.where_object_condition(arel_field, field, value)
+        end.reduce do |condition1, condition2|
+          condition1.and(condition2)
+        end
+
+        where(where_conditions)
       end
 
       def primary_key_is_int?
@@ -96,6 +110,8 @@ module PaperTrail
     # :has_one     set to `false` to opt out of has_one reification.
     #              set to a float to change the lookback time (check whether your db supports
     #              sub-second datetimes if you want them).
+    # :dup         `false` default behavior
+    #              `true` it always create a new object instance. It is useful for comparing two versions of the same object
     def reify(options = {})
       return nil if object.nil?
 
@@ -118,7 +134,7 @@ module PaperTrail
         # `item_type` will be the base class, not the actual subclass.
         # If `type` is present but empty, the class is the base class.
 
-        if item
+        if item && options[:dup] != true
           model = item
           # Look for attributes that exist in the model and not in this version. These attributes should be set to nil.
           (model.attribute_names - attrs.keys).each { |k| attrs[k] = nil }
@@ -193,10 +209,14 @@ module PaperTrail
     end
 
     def index
-      table_name = self.class.table_name
-      @index ||= sibling_versions.
-        select(["#{table_name}.#{PaperTrail.timestamp_field}", "#{table_name}.#{self.class.primary_key}"]).
-        order("#{table_name}.#{PaperTrail.timestamp_field} ASC").index(self)
+      table = self.class.arel_table unless @index
+      @index ||=
+        if self.class.primary_key_is_int?
+          sibling_versions.select(table[self.class.primary_key]).order(table[self.class.primary_key].asc).index(self)
+        else
+          sibling_versions.select([table[PaperTrail.timestamp_field], table[self.class.primary_key]]).
+            order(self.class.timestamp_sort_order).index(self)
+        end
     end
 
     private